Figure out clustering of trees in Tree::analyzeTreeGraph and SkySampleModel::treesInteract.

+++ include/SkySampleModel.h
@@ -63,7 +63,7 @@ public:
   // Member functions - public
   SkySampleModel(float lat);
   ~SkySampleModel(void);
-  unsigned short treesInteract(BoundingBox* B1, BoundingBox* B2);
+  unsigned treesInteract(BoundingBox* B1, BoundingBox* B2);
   bool treesCluster(BoundingBox* B1, BoundingBox* B2);
 
   static SkySampleModel& getSkySampleModel(void) // Get the singleton instance
